#155 Refactoring tests. Fixing pytest fixtures to make test fast again
