+++ tests/integration/it_owner_privilages_test.py
@@ -7,19 +7,19 @@ from timeless.access_control.owner_privileges import has_privilege
 from timeless.employees.models import Employee
 
 
-def test_can_access_location():
+def test_can_access_location(app):
     assert has_privilege(method=Method.CREATE, resource="location")
 
 
-def test_cant_access_unknown_resource():
+def test_cant_access_unknown_resource(app):
     assert not has_privilege(method=Method.CREATE, resource="unknown")
 
 
-def test_cant_access_his_profile():
+def test_cant_access_his_profile(app):
     assert not has_privilege(method=Method.READ, resource="employee", employee_id=1)
 
 
-def test_can_access_his_profile():
+def test_can_access_his_profile(app):
     flask.g.user = Employee(id=1, first_name="Alice", last_name="Cooper",
                       username="alice", phone_number="1",
                       birth_date=datetime.utcnow(),
@@ -28,7 +28,7 @@ def test_can_access_his_profile():
     assert has_privilege(method=Method.READ, resource="employee", employee_id=1)
 
 
-def test_can_access_own_employees():
+def test_can_access_own_employees(app):
     """
     @todo #180:30min We need to clean global object after test finish
      its execution to prevent collision with other tests. Probably we need
